// browser/webapp/src/app/share/models/code-snippet-add-dto.model.ts
import { Language } from './language.model';
import { CodeType } from './code-type.model';
import { Status } from './status.model';
export interface CodeSnippetAddDto {
  /**
   * 实体名称
   */
  name?: string | null;
  /**
   * 描述
   */
  description?: string | null;
  /**
   * 实体定义内容
   */
  content?: string | null;
  /**
   * 语言类型
   */
  language?: Language;
  /**
   * 类型
   */
  codeType?: CodeType;
  /**
   * 状态
   */
  status?: Status;
  updatedTime: Date;
  libraryId?: string | null;

}