// browser/webapp/src/app/share/models/code-type.model.ts
/**
 * 代码片段类型

0 = Function
1 = Util
2 = Service
3 = Controller
4 = Entity
5 = Model
6 = Algorithm
7 = Config
8 = Interface
9 = Else
 */
export enum CodeType {
  Function = 0,
  Util = 1,
  Service = 2,
  Controller = 3,
  Entity = 4,
  Model = 5,
  Algorithm = 6,
  Config = 7,
  Interface = 8,
  Else = 9,

}